Updated .gitignore
Included commonly ignored files for Xcode, Node.JS, Ruby on Rails, IntellijIdea, and Pycharm. Before submitting any code from an IDE not already in our ignore file, we MUST include the ignores in this .gitignore first so that we don't have to remove files from git as its a pain in the ass haha.